JAMES-2997 Move james-server-utils InputStream in a dedicated io package
