+++ server/container/util/src/main/java/org/apache/james/util/io/BodyOffsetInputStream.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations      *
  * under the License.                                           *
  ****************************************************************/
-package org.apache.james.util;
+package org.apache.james.util.io;
 
 import java.io.IOException;
 import java.io.InputStream;
