+++ server/container/util/src/main/java/org/apache/james/util/io/InputStreamUtils.java
@@ -16,27 +16,19 @@
  * specific language governing permissions and limitations      *
  * under the License.                                           *
  ****************************************************************/
-package org.apache.james.util;
+
+package org.apache.james.util.io;
 
 import java.io.InputStream;
+import java.io.SequenceInputStream;
+import java.util.Collections;
 
-public class ZeroedInputStream extends InputStream {
-    public static final int RETURNED_VALUE = 0;
+import com.google.common.collect.Lists;
 
-    private final int max;
-    private int pos;
-
-    public ZeroedInputStream(int max) {
-        this.max = max;
-        this.pos = 0;
-    }
-
-    @Override
-    public int read() {
-        if (pos < max) {
-            pos++;
-            return RETURNED_VALUE;
-        }
-        return -1;
+public class InputStreamUtils {
+    public static InputStream concat(InputStream inputStream, InputStream... additionalInputStreams) {
+        return new SequenceInputStream(
+            Collections.enumeration(
+                Lists.asList(inputStream, additionalInputStreams)));
     }
 }
