+++ server/container/util/src/main/java/org/apache/james/util/io/ZeroedInputStream.java
@@ -6,9 +6,9 @@
  * to you under the Apache License, Version 2.0 (the            *
  * "License"); you may not use this file except in compliance   *
  * with the License.  You may obtain a copy of the License at   *
- * *
- * http://www.apache.org/licenses/LICENSE-2.0                 *
- * *
+ *                                                              *
+ *   http://www.apache.org/licenses/LICENSE-2.0                 *
+ *                                                              *
  * Unless required by applicable law or agreed to in writing,   *
  * software distributed under the License is distributed on an  *
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY       *
@@ -16,29 +16,25 @@
  * specific language governing permissions and limitations      *
  * under the License.                                           *
  ****************************************************************/
+package org.apache.james.util.io;
 
-package org.apache.james.util;
-
-import java.io.IOException;
 import java.io.InputStream;
-import java.util.concurrent.CountDownLatch;
 
-public class CountDownConsumeInputStream extends InputStream {
-    private static final int RETURNED_VALUE = 0;
+public class ZeroedInputStream extends InputStream {
+    public static final int RETURNED_VALUE = 0;
 
-    private final CountDownLatch startSignal;
+    private final int max;
+    private int pos;
 
-    public CountDownConsumeInputStream(CountDownLatch startSignal) {
-        this.startSignal = startSignal;
-    }
-
-    public CountDownLatch getStartSignal() {
-        return startSignal;
+    public ZeroedInputStream(int max) {
+        this.max = max;
+        this.pos = 0;
     }
 
     @Override
-    public int read() throws IOException {
-        if (startSignal.getCount() > 0) {
+    public int read() {
+        if (pos < max) {
+            pos++;
             return RETURNED_VALUE;
         }
         return -1;
